[minor] Updated the readme for readability
**Math rendering**

The math rendering in github is supported by [mathjax](https://www.mathjax.org/).
Updated to support that.
The reason it was not working before (given the comment) is that the mathjax does not expect blank lines after and before `$` or `$$`.

**Link to fork**

Original link was referring to the `master` branch of the parent repository.
This is misleading, as current repository is behind, and is not compatible with the `master` branch anymore.
In addition to that, the default branch name might change, which would cause a broken link.

**Developer guide**

Restructured the developer guide to be more readable.
This will eventually need to be removed to `CONTRIBUTING.md` file.